+++ app/src/main/java/com/cajetan/youtubeplus/MainActivity.kt
@@ -8,8 +8,6 @@ import android.content.Context
 import android.content.Intent
 import android.os.Build
 import android.os.Bundle
-import android.os.PersistableBundle
-import android.support.v7.app.AlertDialog
 import android.support.v7.app.AppCompatActivity
 import android.support.v7.widget.LinearLayoutManager
 import android.util.Log
@@ -20,7 +18,6 @@ import com.cajetan.youtubeplus.data.VideoData
 import com.cajetan.youtubeplus.data.VideoDataViewModel
 import com.cajetan.youtubeplus.utils.YouTubeData
 import com.google.api.services.youtube.model.Video
-import kotlinx.android.synthetic.main.activity_favourites.*
 import kotlinx.android.synthetic.main.activity_main.*
 import org.jetbrains.anko.alert
 import org.jetbrains.anko.intentFor
@@ -33,10 +30,9 @@ class MainActivity : AppCompatActivity(),
 
     private val TAG = this.javaClass.simpleName
 
-    private var mAdapter: VideoListAdapter = VideoListAdapter(emptyList(), this, this)
-    private var mYouTubeData: YouTubeData? = null
-    private var mContext: Context? = null
-    private var mVideoDataViewModel: VideoDataViewModel? = null
+    private val mAdapter: VideoListAdapter = VideoListAdapter(emptyList(), this, this)
+    private lateinit var mYouTubeData: YouTubeData
+    private lateinit var mVideoDataViewModel: VideoDataViewModel
 
     private var mSearchQuery: String = ""
     private var mNextPageToken: String = ""
@@ -51,7 +47,6 @@ class MainActivity : AppCompatActivity(),
         setContentView(R.layout.activity_main)
 
         mYouTubeData = YouTubeData(this)
-        mContext = this
 
         setupSearchResultList()
         createNotificationChannel()
@@ -64,10 +59,7 @@ class MainActivity : AppCompatActivity(),
 
     override fun onResume() {
         super.onResume()
-
-        // TODO: see if the null check is unnecessary
-        if (bottom_bar_main != null)
-            bottom_bar_main.selectedItemId = R.id.action_start
+        bottomBar.selectedItemId = R.id.action_start
     }
 
     override fun onCreateOptionsMenu(menu: Menu?): Boolean {
@@ -86,7 +78,7 @@ class MainActivity : AppCompatActivity(),
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        mYouTubeData?.onParentActivityResult(requestCode, resultCode, data)
+        mYouTubeData.onParentActivityResult(requestCode, resultCode, data)
         super.onActivityResult(requestCode, resultCode, data)
     }
 
@@ -95,7 +87,7 @@ class MainActivity : AppCompatActivity(),
     ////////////////////////////////////////////////////////////////////////////////
 
     private fun setupSearchResultList() {
-        search_results.layoutManager = LinearLayoutManager(this)
+        videoList.layoutManager = LinearLayoutManager(this)
 
         mAdapter.setOnBottomReachedListener {
             Log.d(TAG, "Reached the bottom")
@@ -106,13 +98,13 @@ class MainActivity : AppCompatActivity(),
                 videoSearch(mNextPageToken)
         }
 
-        search_results.setHasFixedSize(false)
-        search_results.adapter = mAdapter
+        videoList.setHasFixedSize(false)
+        videoList.adapter = mAdapter
     }
 
     private fun setupBottomBar() {
-        bottom_bar_main.selectedItemId = R.id.action_start
-        bottom_bar_main.setOnNavigationItemSelectedListener {
+        bottomBar.selectedItemId = R.id.action_start
+        bottomBar.setOnNavigationItemSelectedListener {
              when (it.itemId) {
                 R.id.action_start -> {
                     it.setChecked(true)
@@ -158,25 +150,25 @@ class MainActivity : AppCompatActivity(),
         }
 
         if (nextPageToken == null) {
-            search_progress_bar_centre.visibility = View.VISIBLE
-            search_results.visibility = View.INVISIBLE
+            searchProgressBarCentre.visibility = View.VISIBLE
+            videoList.visibility = View.INVISIBLE
         } else {
-            search_progress_bar_bottom.visibility = View.VISIBLE
+            searchProgressBarBottom.visibility = View.VISIBLE
         }
 
         Log.d(TAG, "Searching for a video $mSearchQuery")
-        mYouTubeData?.receiveSearchResults(mSearchQuery, nextPageToken)
+        mYouTubeData.receiveSearchResults(mSearchQuery, nextPageToken)
     }
 
     private fun loadMostPopularVideos(nextPageToken: String?) {
         if (nextPageToken == null) {
-            search_progress_bar_centre.visibility = View.VISIBLE
-            search_results.visibility = View.INVISIBLE
+            searchProgressBarCentre.visibility = View.VISIBLE
+            videoList.visibility = View.INVISIBLE
         } else {
-            search_progress_bar_bottom.visibility = View.VISIBLE
+            searchProgressBarBottom.visibility = View.VISIBLE
         }
 
-        mYouTubeData?.receiveMostPopularResults(nextPageToken)
+        mYouTubeData.receiveMostPopularResults(nextPageToken)
     }
 
     private fun handleIntent(intent: Intent) {
@@ -199,12 +191,12 @@ class MainActivity : AppCompatActivity(),
                                          nextPageToken: String?, previousPageToken: String?) {
         if (previousPageToken == null || previousPageToken == "") {
             mAdapter.clearItems()
-            search_results.scrollToPosition(0)
+            videoList.scrollToPosition(0)
 
-            search_progress_bar_centre.visibility = View.INVISIBLE
-            search_results.visibility = View.VISIBLE
+            searchProgressBarCentre.visibility = View.INVISIBLE
+            videoList.visibility = View.VISIBLE
         } else {
-            search_progress_bar_bottom.visibility = View.GONE
+            searchProgressBarBottom.visibility = View.GONE
         }
 
         mAdapter.addItems(results)
@@ -215,12 +207,12 @@ class MainActivity : AppCompatActivity(),
                                        nextPageToken: String?, previousPageToken: String?) {
         if (previousPageToken == null || previousPageToken == "") {
             mAdapter.clearItems()
-            search_results.scrollToPosition(0)
+            videoList.scrollToPosition(0)
 
-            search_progress_bar_centre.visibility = View.INVISIBLE
-            search_results.visibility = View.VISIBLE
+            searchProgressBarCentre.visibility = View.INVISIBLE
+            videoList.visibility = View.VISIBLE
         } else {
-            search_progress_bar_bottom.visibility = View.GONE
+            searchProgressBarBottom.visibility = View.GONE
         }
 
         mAdapter.addItems(results)
@@ -235,7 +227,7 @@ class MainActivity : AppCompatActivity(),
 
     override fun onListItemLongClick(clickedVideoId: String?) {
         alert(getString(R.string.favourite_add_confirmation)) {
-            yesButton { mVideoDataViewModel?.insert(VideoData(clickedVideoId as String)) }
+            yesButton { mVideoDataViewModel.insert(VideoData(clickedVideoId as String)) }
             noButton { }
         }.show()
     }